poke_team/urls.py
from django.contrib import admin
from django.urls import path, include
from django.conf.urls import url
from drf_yasg.views import get_schema_view
from drf_yasg import openapi
from rest_framework import permissions

app_name = "core"

schema_view = get_schema_view(
    openapi.Info(
       title="PokeTeam",
       default_version='v1',
       description="Pokemon team management API.",
       contact=openapi.Contact(email="dionvictor11@gmail.com"),
    ),
    public=True,
    permission_classes=(permissions.AllowAny,),
)

urlpatterns = [
    path('admin/', admin.site.urls),
    path('accounts/', include('rest_framework.urls')),
    path('', include('core.urls'), name='core'),
    url(r'^swagger/$', schema_view.with_ui('swagger', cache_timeout=0), name='schema-swagger-ui'),
]
