requirements.txt
Django==3.2.5
psycopg2==2.8.6
psycopg2-binary==2.8.6
djangorestframework==3.12.4
drf_yasg==1.20.0
